requirements.txt
behave==1.2.6
Faker==4.0.0
parse==1.14.0
parse-type==0.5.2
PyHamcrest==2.0.0
python-dateutil==2.8.1
selenium==3.141.0
six==1.14.0
text-unidecode==1.3
urllib3==1.25.8
